update definitions of hist_data and stn_datasets

+++ scripts/stn_dataset.hh
@@ -8,8 +8,8 @@ struct stn_dataset_t {
 
   TString  fID        ;			// dataset ID (short name)
   TString  fName      ;			// full name of the ntupled dataset
-  int      fNEvents   ;                 // number of events in all files
-  int      fNGenEvents; 		// N generated events (if MC, corresponding to the files, for normalization)
+  long int fNEvents   ;                 // number of events in all files
+  long int fNGenEvents; 		// N generated events (if MC, corresponding to the files, for normalization)
   float    fEMin      ;			// sometimes, the signal particle energy range is constrained
   float    fEMax      ;
   int      fMinRun    ;                 // minimal run number
